+++ src/index.ts
@@ -1,3 +1,7 @@
 export { MaraDocsClient } from "./MaraDocsClient/index";
 export { MaraDocsServer } from "./MaraDocsServer/index";
 export type { RequestOptions } from "./shared/requestOptions";
+export type {
+  CreateDownloadProxyResult,
+  CreateUploadResult,
+} from "./MaraDocsClient/dataEp";